Revert "double formatting in FieldSolver.cu"
This reverts commit 87ec54bd8f98e7f848f47b140968f6416b833e20.

+++ FieldSolver.cu
@@ -84,7 +84,7 @@ __global__ void EvaluateFields(const double* dev_potential, double3* dev_el_fiel
 	is_on_low_border = (blockIdx.x == (gridDim.x - 1));
 	is_inside_borders = !(is_on_low_border || is_on_up_border);
 
-	e.x = -(1.0 / (1.0 + is_inside_borders)) * GradientComponent(
+	e.x = -((double)1 / ((double)1 + is_inside_borders)) * GradientComponent(
 		dev_potential[idx + (offset*is_on_up_border) - (offset*is_inside_borders)],
 		dev_potential[idx - (offset*is_on_low_border) + (offset*is_inside_borders)],
 		d_cell_size[0].x);
@@ -94,7 +94,7 @@ __global__ void EvaluateFields(const double* dev_potential, double3* dev_el_fiel
 	is_on_low_border = (blockIdx.y == (gridDim.y - 1));
 	is_inside_borders = !(is_on_low_border || is_on_up_border);
 
-	e.y = -(1.0 / (1.0 + is_inside_borders)) * GradientComponent(
+	e.y = -((double)1 / ((double)1 + is_inside_borders)) * GradientComponent(
 		dev_potential[idx + (offset*is_on_up_border) - (offset*is_inside_borders)],
 		dev_potential[idx - (offset*is_on_low_border) + (offset*is_inside_borders)],
 		d_cell_size[0].y);
@@ -104,7 +104,7 @@ __global__ void EvaluateFields(const double* dev_potential, double3* dev_el_fiel
 	is_on_low_border = (blockIdx.z == (gridDim.z - 1));
 	is_inside_borders = !(is_on_low_border || is_on_up_border);
 
-	e.z = -(1.0 / (1.0 + is_inside_borders)) * GradientComponent(
+	e.z = -((double)1 / ((double)1 + is_inside_borders)) * GradientComponent(
 		dev_potential[idx + (offset * is_on_up_border) - (offset * is_inside_borders)],
 		dev_potential[idx - (offset * is_on_low_border) + (offset * is_inside_borders)],
 		d_cell_size[0].z);
